libs/python/shared/py-dotenv/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "pydotenv" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 --cov-report html:'../../../../coverage/libs/python/shared/py-dotenv/html' --cov-report xml:'../../../../coverage/libs/python/shared/py-dotenv/coverage.xml' --html='../../../../reports/libs/python/shared/py-dotenv/unittests/html/index.html' --junitxml='../../../../reports/libs/python/shared/py-dotenv/unittests/junit.xml'"

[tool.poetry]
name = "shared-py-dotenv"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "pydotenv"

  [tool.poetry.dependencies]
  python = ">=3.9,<4.0"
  python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
